refactor: convert jazzicon.component.js to TypeScript
