+++ ui/components/ui/jazzicon/jazzicon.d.ts
@@ -1,0 +1,4 @@
+declare module '@metamask/jazzicon' {
+  function jazzicon(diameter: number, seed: number): SVGSVGElement;
+  export default jazzicon;
+}
